// src/api/error/ErrorCode.ts
export enum ErrorCode {
    UNKNOWN_ERROR_CODE = "NLN-UNK",
    MOVIES_NOT_FOUND_CODE = "NLN-0002",
    MOVIE_NOT_FOUND_CODE = "NLN-0003",
    ROOMS_NOT_FOUND_CODE = "NLN-0004",
    ROOM_NOT_FOUND_CODE = "NLN-0005",
    ROOM_ALREADY_EXISTS_CODE = "NLN-0006",
    SESSIONS_NOT_FOUND_CODE = "NLN-0007",
    SESSION_NOT_FOUND_CODE = "NLN-0008",
    SESSION_BY_ROOM_NOT_FOUND_CODE = "NLN-0009",
    SESSION_IN_GIVEN_TIME_ALREADY_EXISTS_CODE = "NLN-0010",
    PARAMETER_NULL_CODE = "NLN-0011",
    PARAMETER_INVALID_CODE = "NLN-0012",
    PARAMETER_ARRAY_SHORT_CODE = "NLN-0013",
    PARAMETER_INVALID_SIZE_CODE = "NLN-0014",
    VERIFICATION_INVALID_CODE = "NLN-0015",
    USER_NOT_FOUND_CODE = "NLN-0016",
    PASSWORD_INCORRECT_CODE = "NLN-0017",
    USER_ALREADY_REGISTERED_CODE = "NLN-0018",
    NOT_AUTHENTICATED_CODE = "NLN-0019",
    MISSING_AUTH_CODE = "NLN-0021",
    DATE_ISO_REQUIRED_CODE = "NLN-0022",
    MOVIE_NAME_TOO_LARGE_CODE = "NLN-0023",
    MOVIE_WRONG_NAME_CODE = "NLN-0024",
    GLOBALPARAMETER_NOT_FOUND_CODE = "NLN-0025",
    GLOBALPARAMETER_ALREADY_EXISTS_CODE = "NLN-0026",
    ORIGIN_NOT_ALLOWED_CODE = "NLN-0027",
}